refactor: add styles folder and move styles file into
