+++ index.html
@@ -5,7 +5,7 @@
     <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Rick and Morty</title>
-     <link href="/src/App.css" rel="stylesheet">
+     <link href="src/assets/styles/App.css" rel="stylesheet">
   </head>
   <body>
     <div id="root"></div>
